Code organization according to The Official raywenderlich.com Swift Style Guide. Split code into files QuizStepViewModel, QuizQuestion, QuizResultsViewModel.
