+++ MovieQuiz/Presentation/QuizResultsViewModel.swift
@@ -1,0 +1,18 @@
+//
+//  QuizResultsViewModel.swift
+//  MovieQuiz
+//
+//  Created by ulyana on 8.10.24.
+//
+
+import Foundation
+
+// для состояния "Результат квиза"
+struct QuizResultsViewModel {
+    // строка с заголовком алерта
+    let title: String
+    // строка с текстом о количестве набранных очков
+    let text: String
+    // текст для кнопки алерта
+    let buttonText: String
+}
